+++ src/lib/domains/global/index.ts
@@ -4,14 +4,14 @@ import { persist } from "zustand/middleware";
 
 import { IS_DEV } from "@/config";
 import { type ColorSchemeSliceType, createColorSchemeSlice } from "@/lib/domains/colorScheme/store";
-import { type FoobarSliceType, createFoobarSlice } from "@/lib/domains/foobar/flags";
+import { type FoobarSliceType, createFoobarSlice } from "@/lib/domains/foobar/store";
 
 type CombinedState = FoobarSliceType & ColorSchemeSliceType;
 export const useGlobalStore = create<CombinedState>()(
 	persist((...a) => ({ ...createFoobarSlice(...a), ...createColorSchemeSlice(...a) }), {
 		name: IS_DEV ? "foobar-zustand-dev" : "foobar-zustand",
 		partialize: (state) => ({
-			foobarData: state.foobarData,
+			foobar_data: state.foobar_data,
 		}),
 		merge: (persistedState, currentState) => merge(currentState, persistedState),
 		onRehydrateStorage: () => (state) => {
